[PUBDEV-9048] fix precission in tests

+++ h2o-py/tests/pyunit_utils/utilsPY.py
@@ -1140,7 +1140,7 @@ def generate_one_cluster(cluster_center, cluster_number, cluster_size):
 
     pt_dists = np.random.uniform(0, cluster_size, [cluster_number, 1])
     coord_pts = len(cluster_center)     # dimension of each cluster point
-    one_cluster_data = np.zeros((cluster_number, coord_pts), dtype=np.float64)
+    one_cluster_data = np.zeros((cluster_number, coord_pts), dtype=np.float32)
 
     for p_ind in range(cluster_number):
         coord_indices = list(range(coord_pts))
@@ -1625,7 +1625,7 @@ def duplicate_scale_cols(col_indices, col_scale, old_filename, new_filename):
 
     np_frame = np.asmatrix(np.genfromtxt(old_filename, delimiter=',', dtype=None))
     (num_row, num_col) = np_frame.shape
-    np_frame_new = np.asmatrix(np.zeros((num_row, len(col_indices)), dtype=np.float64))
+    np_frame_new = np.asmatrix(np.zeros((num_row, len(col_indices)), dtype=np.float32))
 
     for ind in range(len(col_indices)):
         np_frame_new[:, ind] = np_frame[:, col_indices[ind]]*col_scale[ind]
@@ -2067,7 +2067,7 @@ def get_train_glm_params(model, what_param, family_type='gaussian'):
             num_feature = len(coeff_pvalues)
             num_class = (len(coeff_pvalues[0])-1)/2
 
-            coeffs = np.zeros((num_class,num_feature), dtype=np.float64)
+            coeffs = np.zeros((num_class,num_feature), dtype=np.float32)
 
             end_index = int(num_class+1)
             for col_index in range(len(coeff_pvalues)):
